erro factory

+++ LibraryApi/src/main/java/com/LibraryApi/model/Livro.java
@@ -4,6 +4,7 @@ package com.LibraryApi.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -29,7 +30,8 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale =2 )
-    private Double preco;
+    private BigDecimal preco;
+
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
